Roadmap_ks to Roadmap

See merge request s04-webmobile1-sub2/s04p12c106!6

+++ backend/src/main/java/com/web/blog/model/service/FreeBoardServiceImpl.java
@@ -6,7 +6,9 @@ import java.util.Map;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestBody;
 
+import com.web.blog.model.dto.Comment;
 import com.web.blog.model.dto.Posting;
 import com.web.blog.model.repo.CommentRepo;
 import com.web.blog.model.repo.PostingRepo;
@@ -112,6 +114,7 @@ public class FreeBoardServiceImpl implements FreeBoardService{
 		Map<String, Object> result = new HashMap<String, Object>();
 		try {
 			result.put("posting", postingRepo.select(Integer.parseInt(pid)));
+			result.put("comments", commentRepo.selectListPid(Integer.parseInt(pid)));
 			result.put("msg", "success");
 		} catch(NumberFormatException e){
 			logger.error("input data type error");
@@ -123,7 +126,7 @@ public class FreeBoardServiceImpl implements FreeBoardService{
 		return result;
 	}
 	
-	public Object regist(Posting posting) {
+	public Object registPosting(Posting posting) {
 		Map<String, Object> result = new HashMap<String, Object>();
 		try {
 			if(postingRepo.insert(posting) == 1) result.put("msg", "success");
@@ -138,9 +141,10 @@ public class FreeBoardServiceImpl implements FreeBoardService{
 		return result;
 	}
 	
-	public Object edit(Posting posting) {
+	public Object editPosting(Posting posting, String uid) {
 		Map<String, Object> result = new HashMap<String, Object>();
 		try {
+			if(postingRepo.select(posting.getPid()).getUid() != Integer.parseInt(uid)) throw new RuntimeException("wrong user");
 			if(postingRepo.update(posting) == 1) result.put("msg", "success");
 			else result.put("msg", "fail");
 		} catch(NumberFormatException e){
@@ -153,10 +157,58 @@ public class FreeBoardServiceImpl implements FreeBoardService{
 		return result;
 	}
 	
-	public Object delete(String pid) {
+	public Object deletePosting(String pid, String uid) {
 		Map<String, Object> result = new HashMap<String, Object>();
 		try {
+			if(postingRepo.select(Integer.parseInt(pid)).getUid() != Integer.parseInt(uid)) throw new RuntimeException("wrong user");
 			if(postingRepo.delete(Integer.parseInt(pid)) == 1) result.put("msg", "success");
+			else result.put("msg", "fail");
+		} catch(NumberFormatException e){
+			logger.error("input data type error");
+			result.put("msg", "fail");
+		} catch(Exception e) {
+			logger.error("i dont know");
+			result.put("msg", "fail");
+		}
+		return result;
+	}
+	
+	public Object registComment(Comment comment) {
+		Map<String, Object> result = new HashMap<String, Object>();
+		try {
+			if(commentRepo.insert(comment) == 1) result.put("msg", "success");
+			else result.put("msg", "fail");
+		} catch(NumberFormatException e){
+			logger.error("input data type error");
+			result.put("msg", "fail");
+		} catch(Exception e) {
+			logger.error("i dont know");
+			result.put("msg", "fail");
+		}
+		return result;
+	}
+	
+	public Object editComment(Comment comment, String uid) {
+		Map<String, Object> result = new HashMap<String, Object>();
+		try {
+			if(commentRepo.select(comment.getCid()).getUid() != Integer.parseInt(uid)) throw new RuntimeException("wrong user");
+			if(commentRepo.update(comment) == 1) result.put("msg", "success");
+			else result.put("msg", "fail");
+		} catch(NumberFormatException e){
+			logger.error("input data type error");
+			result.put("msg", "fail");
+		} catch(Exception e) {
+			logger.error("i dont know");
+			result.put("msg", "fail");
+		}
+		return result;
+	}
+	
+	public Object deleteComment(String cid, String uid) {
+		Map<String, Object> result = new HashMap<String, Object>();
+		try {
+			if(commentRepo.select(Integer.parseInt(cid)).getUid() != Integer.parseInt(uid)) throw new RuntimeException("wrong user");
+			if(commentRepo.delete(Integer.parseInt(cid)) == 1) result.put("msg", "success");
 			else result.put("msg", "fail");
 		} catch(NumberFormatException e){
 			logger.error("input data type error");
